==> src/notebooks/01_json_scraping.ipynb <==
import json

file_path = '../../data/json_handling/spotify_playlist.json'

with open(file_path, 'r') as f:
    data = json.load(f)

for track in data['contents']['items']:
    print(track['playCount'])
    print(track['name'])  
    print()
