==> src/notebooks/02_web_scraping.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import requests

response = requests.get(url="https://en.wikipedia.org/wiki/Community_areas_in_Chicago")

soup = BeautifulSoup(response.content, 'html.parser')

title = soup.find(id="firstHeading")
print(title.string)

for table in soup.find_all('table'):
    print(table.get("class"))

table = soup.find('table', class_='wikitable sortable plainrowheaders mw-datatable')
for row in table.tbody.find_all('tr')[2:-1]:
    print(row)
    print("----")

for row in table.tbody.find_all('tr')[2:-1]:
    cells = row.find_all('td')
    print(cells[0].get_text(strip=True))

for row in table.tbody.find_all('tr')[2:-1]:
    header_cell = row.find('th')
    print(header_cell.a.get_text(strip=True))

data = []

for row in table.tbody.find_all('tr')[2:-1]:
    cells = row.find_all('td')
    header_cell = row.find('th')
    area_codes = cells[0].get_text(strip=True)
    community_name = header_cell.a.get_text(strip=True)
    # print(area_codes, community_name)

    data.append({"area_codes": area_codes, "community_name": community_name})

data

community_areas = pd.DataFrame(data)
# check Dtype, etc.
# community_areas.info()
community_areas['area_codes'] = community_areas['area_codes'].astype(int)

community_areas
community_areas.info()

community_areas.to_csv("community_areas_master.csv", index=False)
