==> src/notebooks/07_transform_load.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta

import pandas as pd
pd.set_option('display.max_columns', 30)

import requests

"""
1. get the data from S3
2. weather data transformations
3. taxi_trips transformations
4. update payment_type master
5. update company master
6. update taxi_trips with payment_type and company ids (replace the string values with the ids from the latest master tables)
7. upload weather data to S3
8. upload taxi data to S3
9. upload the newest payment_type and company master tables to S3
"""

#### taxi_trips transformation codes

current_datetime = datetime.now() - relativedelta(months=2)

formatted_date = current_datetime.strftime("%Y-%m-%d")

url = f"https://data.cityofchicago.org/resource/ajtu-isnz.json?$where=trip_start_timestamp >= '{formatted_date}T00:00:00' AND trip_start_timestamp <= '{formatted_date}T23:59:59'&$limit=30000"

response = requests.get(url)

data = response.json()

taxi_trips = pd.DataFrame(data)

taxi_trips.drop(["pickup_census_tract", "dropoff_census_tract"], axis=1, inplace=True)
taxi_trips.drop(["pickup_centroid_location", "dropoff_centroid_location"], axis=1, inplace=True)

taxi_trips.dropna(inplace=True)

taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
                           "dropoff_community_area": "dropoff_community_area_id"}, inplace=True)

taxi_trips["datetime_for_weather"] = pd.to_datetime(taxi_trips["trip_start_timestamp"]).dt.floor("h")

taxi_trips.head()

#### taxi_trips transformation function

def taxi_trips_transformations(taxi_trips: pd.DataFrame) -> pd.DataFrame:
    """
    Perform transformations on the taxi_trips DataFrame

    Parameters
    ---------- 
    taxi_trips: pd.DataFrame
        The DataFrame containing the taxi trips data

    Returns
    -------
    pd.DataFrame
        The transformed DataFrame
    """

    if not isinstance(taxi_trips, pd.DataFrame):
        raise TypeError("taxi_trips should be a DataFrame")

    taxi_trips.drop(["pickup_census_tract", "dropoff_census_tract", 
                     "pickup_centroid_location", "dropoff_centroid_location"], axis=1, inplace=True)

    taxi_trips.dropna(inplace=True)

    taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
                               "dropoff_community_area": "dropoff_community_area_id"}, inplace=True)

    taxi_trips["datetime_for_weather"] = pd.to_datetime(taxi_trips["trip_start_timestamp"]).dt.floor("h")

    return taxi_trips

#### company update codes

company_master = taxi_trips["company"].drop_duplicates().reset_index(drop=True)
company_master = pd.DataFrame({
    "company_id": range(1, len(company_master) + 1),
    "company": company_master
})
company_master.tail()

new_company_data = [
    { "company": "6574 - Babylon Express Inc." },
    { "company": "Y" },
    { "company": "X" }
]

new_company_mapping = pd.DataFrame(new_company_data)
new_company_mapping

company_max_id = company_master["company_id"].max()
company_max_id

new_companies_list = []

# one line
new_companies_list = [company for company in new_company_mapping["company"].values if company not in company_master["company"].values]

new_companies_list

new_companies_df = pd.DataFrame({
    "company_id": range(company_max_id + 1, company_max_id + 1 + len(new_companies_list)),
    "company": new_companies_list
})
new_companies_df

updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)
updated_company_master.tail()

def update_company_master(taxi_trips: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    """
    Update the company_master DataFrame with new companies from the taxi_trips DataFrame
    
    Paramters
    ---------
    taxi_trips: pd.DataFrame
        The DataFrame containing the taxi trips data
    company_master: pd.DataFrame
        The DataFrame containing the company master data

    Returns
    -------
    pd.DataFrame
        The updated company_master DataFrame
    """
    company_max_id = company_master["company_id"].max()

    new_companies_list = [company for company in taxi_trips["company"].values if company not in company_master["company"].values]

    new_companies_df = pd.DataFrame({
        "company_id": range(company_max_id + 1, company_max_id + 1 + len(new_companies_list)),
        "company": new_companies_list
    })

    updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)

    return updated_company_master

updated_company_master = update_company_master(taxi_trips, company_master)
updated_company_master.tail()

taxi_trips_company_only = pd.DataFrame({
    "company_id": [1,2,3],
    "company": ["6574 - Babylon Express Inc.", "Y", "X"]
})
taxi_trips_company_only

### payment_type_master codes

payment_type_master = taxi_trips["payment_type"].drop_duplicates().reset_index(drop=True)
payment_type_master = pd.DataFrame({
    "payment_type_id": range(1, len(payment_type_master) + 1),
    "payment_type": payment_type_master
})

taxi_trips_payment_type_only = pd.DataFrame({
    "payment_type_id": [1,2,3],
    "payment_type": ["Credit Card", "Y", "X"]
})
taxi_trips_payment_type_only

def update_payment_type_master(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame) -> pd.DataFrame:
    """
    Update the payment_type_master DataFrame with new companies from the taxi_trips DataFrame
    
    Paramters
    ---------
    taxi_trips: pd.DataFrame
        The DataFrame containing the taxi trips data
    payment_type_master: pd.DataFrame
        The DataFrame containing the payment_type master data

    Returns
    -------
    pd.DataFrame
        The updated payment_type_master DataFrame
    """
    payment_type_max_id = payment_type_master["payment_type_id"].max()

    new_payment_types_list = [payment_type for payment_type in taxi_trips["payment_type"].values if payment_type not in payment_type_master["payment_type"].values]

    new_payment_types_df = pd.DataFrame({
        "payment_type_id": range(payment_type_max_id + 1, payment_type_max_id + 1 + len(new_payment_types_list)),
        "payment_type": new_payment_types_list
    })

    updated_payment_type_master = pd.concat([payment_type_master, new_payment_types_df], ignore_index=True)

    return updated_payment_type_master

updated_payment_type_master = update_payment_type_master(taxi_trips_payment_type_only, payment_type_master)
updated_payment_type_master.tail()

### Create a generic update master function

def update_master(taxi_trips: pd.DataFrame, master: pd.DataFrame, id_column: str, value_column: str) -> pd.DataFrame:
    """
    Update the master DataFrame with new companies from the taxi_trips DataFrame
    
    Paramters
    ---------
    taxi_trips: pd.DataFrame
        The DataFrame containing the taxi trips data
    master: pd.DataFrame
        The DataFrame containing the master data
    id_column: str
        The id column of the master DataFrame.
    value_column: str
        The name of the column in master_df containing the values.

    Returns
    -------
    pd.DataFrame
        The updated master DataFrame
    """
    max_id = master[id_column].max()

    new_values_list = list(set(taxi_trips[value_column].values) - set(master[value_column].values))
    new_values_df = pd.DataFrame({
        id_column: range(max_id + 1, max_id + 1 + len(new_values_list)),
        value_column: new_values_list
    })

    updated_master = pd.concat([master, new_values_df], ignore_index=True)

    return updated_master

# test_payment_type_master = update_master(taxi_trips_payment_type_only, payment_type_master, "payment_type_id", "payment_type")
# test_company_master = update_master(taxi_trips_company_only, company_master, "company_id", "company")

#### Updating taxi_trips with the most recent company_master and payment_type master codes

taxi_trips.head()

def update_taxi_trips_with_master_data(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    '''
    Update the taxi_trips DataFrame with the payment_type_master and company_master ids

    Parameters
    ----------
    taxi_trips: pd.DataFrame
        The DataFrame containing the taxi trips data
    payment_type_master: pd.DataFrame
        The DataFrame containing the payment_type master table
    company_master: pd.DataFrame
        The DataFrame containing the company master table

    Returns
    -------
    pd.DataFrame
        The updated taxi_trips DataFrame with the payment_type and company ids, without the string values
    '''
    taxi_trips_id  = taxi_trips.merge(payment_type_master, on="payment_type")
    taxi_trips_id  = taxi_trips_id.merge(company_master, on="company")
    
    taxi_trips_id.drop(["payment_type", "company"], axis=1, inplace=True)
    
    return taxi_trips_id

taxi_trips_id = update_taxi_trips_with_master_data(taxi_trips, updated_payment_type_master, updated_company_master)
taxi_trips_id.sample(5)

### weather transformation function

def transform_weather_data(weather_data: JSON) -> pd.DataFrame:
    """
    Perform transformations on the weather data
    
    Parameters
    ----------
    weather_data: dict
        The daily weather data from the Open Meteo API

    Returns
    -------
    pd.DataFrame
        The transformed weather data DataFrame
    """
    weather_data_filtered = {
        "datetime": weather_data['hourly']['time'],
        "temperature": weather_data['hourly']['temperature_2m'],
        "wind_speed": weather_data['hourly']['wind_speed_10m'],
        "rain": weather_data['hourly']['rain'],
        "precipitation": weather_data['hourly']['precipitation']
    }

    weather_df = pd.DataFrame(weather_data_filtered)

    weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])

    return weather_df

# Get weather data and create a dataframe out of it
current_datetime = datetime.now() - relativedelta(months=2)
formatted_date = current_datetime.strftime("%Y-%m-%d")

url = 'https://archive-api.open-meteo.com/v1/era5'

params = {
    'start_date': formatted_date,
    'end_date': formatted_date,
    'latitude': 41.85,
    'longitude': -87.65,
    "hourly": "temperature_2m,wind_speed_10m,rain,precipitation",
}

response = requests.get(url, params=params)
weather_data = response.json()

transform_weather_data(weather_data).tail()
