==> src/notebooks/08_local_visualizations.ipynb <==
import os
import boto3
import pandas as pd
from io import StringIO

pd.set_option('display.max_columns', 50)

aws_access_key_id = os.getenv("AWS_ACCESS_KEY")
aws_secret_key = os.getenv("AWS_SECRET_KEY")

def read_csv_from_s3(bucket: str, path: str, filename: str) -> pd.DataFrame:
    """
    Read a CSV file from S3 and return it as a DataFrame

    Parameters
    ----------
    bucket: str
        The name of the S3 bucket
    path: str
        The path to the CSV file in the S3 bucket
    filename: str
        The name of the CSV file

    Returns
    -------
    pd.DataFrame
        The DataFrame containing the CSV data
    """
    s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_key)
    obj = s3.get_object(Bucket=bucket, Key=f"{path}{filename}")

    output_df = pd.read_csv(StringIO(obj['Body'].read().decode('utf-8')))

    return output_df

### Prepare folders

s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_key)
bucket = 'cubix-chicago-taxi-34'

community_areas_path = "transformed_data/community_areas/"
company_path = "transformed_data/company/"
date_path = "transformed_data/date/"
payment_type_path = "transformed_data/payment_type/"
taxi_trips_path = "transformed_data/taxi_trips/"
weather_path = "transformed_data/weather/"

### Read files

community_areas = read_csv_from_s3(bucket, community_areas_path, "community_areas_master.csv")
company = read_csv_from_s3(bucket, company_path, "company_master.csv")
date = read_csv_from_s3(bucket, date_path, "date_dimensions.csv")
payment_type = read_csv_from_s3(bucket, payment_type_path, "payment_type_master.csv")

trips_list = []
weather_list = []

for file in s3.list_objects_v2(Bucket=bucket, Prefix=taxi_trips_path)['Contents']:
    file_key = file['Key']

    if(file_key.split('/'))[-1].strip() != "":
        if(file_key.split('.')[1] == "csv"):
            filename = file_key.split('/')[-1]

            trip = read_csv_from_s3(bucket, taxi_trips_path, filename)
       
            trips_list.append(trip)
            print(f"{filename} has been added")

trips = pd.concat(trips_list, ignore_index=True)

for file in s3.list_objects_v2(Bucket=bucket, Prefix=weather_path)['Contents']:
    file_key = file['Key']

    if(file_key.split('/'))[-1].strip() != "":
        if(file_key.split('.')[1] == "csv"):
            filename = file_key.split('/')[-1]

            weather_daily = read_csv_from_s3(bucket, weather_path, filename)
            weather_list.append(weather_daily)
            print(f"{filename} has been added")

weather = pd.concat(weather_list, ignore_index=True)

### Join data together

trips_full = pd.merge(trips, weather, how='inner', left_on='datetime_for_weather', right_on='datetime')
trips_full = trips_full.drop(columns=['datetime'])

trips_full = pd.merge(trips_full, company, how='inner', left_on='company_id', right_on='company_id')
trips_full = trips_full.drop(columns=['company_id'])

trips_full = pd.merge(trips_full, payment_type, how='inner', left_on='payment_type_id', right_on='payment_type_id')
trips_full = trips_full.drop(columns=['payment_type_id'])

trips_full = pd.merge(trips_full, community_areas, how='inner', left_on='pickup_community_area_id', right_on='area_codes')
trips_full = trips_full.drop(columns=['pickup_community_area_id', 'area_codes'])
trips_full.rename(columns={'community_name': 'pickup_community_area_name'}, inplace=True)

trips_full = pd.merge(trips_full, community_areas, how='inner', left_on='dropoff_community_area_id', right_on='area_codes')
trips_full = trips_full.drop(columns=['dropoff_community_area_id', 'area_codes'])
trips_full.rename(columns={'community_name': 'dropoff_community_area_name'}, inplace=True)

date['date'] = pd.to_datetime(date['date'])
trips_full['trip_start_timestamp'] = pd.to_datetime(trips_full['trip_start_timestamp'])
trips_full['trip_start_date'] = trips_full['trip_start_timestamp'].dt.date
trips_full['trip_start_date'] = pd.to_datetime(trips_full['trip_start_date'])

trips_full = pd.merge(trips_full, date, how='inner', left_on='trip_start_date', right_on='date')
trips_full = trips_full.drop(columns=['date'])

trips_full.head()
